refactor: separat Icon stories
