+++ packages/admin-stories/src/docs/comet/icons/IconList.stories.mdx
@@ -6,4 +6,4 @@ import { Meta, Canvas, Story } from "@storybook/addon-docs/blocks";
 
 ## Icon List
 
-<Story id="stories-icons--icon-list" />
+<Story id="stories-icons-list--icon-list" />
