+++ packages/admin-stories/src/docs/comet/icons/IconsUsage.stories.mdx
@@ -13,32 +13,32 @@ import { Attachment } from "@comet/admin-icons";
 ```
 
 <Canvas>
-    <Story id="stories-icons--icon" />
+    <Story id="stories-icons-render-icon--icon" />
 </Canvas>
 
 ## Sizes
 
 <Canvas>
-    <Story id="stories-icons--small-size-icon" />
-    <Story id="stories-icons--default-size-icon" />
-    <Story id="stories-icons--large-size-icon" />
-    <Story id="stories-icons--custom-size-icon" />
+    <Story id="stories-icons-sizes--small-size-icon" />
+    <Story id="stories-icons-sizes--default-size-icon" />
+    <Story id="stories-icons-sizes--large-size-icon" />
+    <Story id="stories-icons-sizes--custom-size-icon" />
 </Canvas>
 
 ## Render icon in text
 
 <Canvas>
-    <Story id="stories-icons--render-icon-in-text" />
+    <Story id="stories-icons-render-icon-in-text--render-icon-in-text" />
 </Canvas>
 
 ## Colors
 
 <Canvas>
-    <Story id="stories-icons--default-color" />
-    <Story id="stories-icons--primary-color" />
-    <Story id="stories-icons--secondary-color" />
-    <Story id="stories-icons--error-color" />
-    <Story id="stories-icons--disabled-color" />
-    <Story id="stories-icons--action-color" />
-    <Story id="stories-icons--custom-color" />
+    <Story id="stories-icons-colors--default-color" />
+    <Story id="stories-icons-colors--primary-color" />
+    <Story id="stories-icons-colors--secondary-color" />
+    <Story id="stories-icons-colors--error-color" />
+    <Story id="stories-icons-colors--disabled-color" />
+    <Story id="stories-icons-colors--action-color" />
+    <Story id="stories-icons-colors--custom-color" />
 </Canvas>
