+++ packages/admin-stories/src/docs/comet/icons/stories/RenderIcon.stories.tsx
@@ -1,0 +1,7 @@
+import { Attachment } from "@comet/admin-icons";
+import { storiesOf } from "@storybook/react";
+import * as React from "react";
+
+storiesOf("stories/Icons/Render Icon", module).add("Icon", () => {
+    return <Attachment />;
+});
